fix(showcases): fix devoxx video (#247)
